Some changes
- Fix some compilation issues on windows
- Update pack.sh to support bash on windows

+++ pack.sh
@@ -1,32 +1,58 @@
 #!/bin/bash
 
-BUILD_DIR="build/android"
+# format:
+# pack.sh -v 15b3 -m "Custom description (optional)"
+# if using bash on windows, do:
+# pack.sh win -v 15b3 -m "Custom description" -p windows
+# pack.sh win -v 15b3 -m "Custom description" -p android
+# requires sed(for editing), zip(for packing)
+
 BUILD_SCRIPT="./build.sh"
 PACK_DIR="pack"
-TEMP_PACK_DIR="$BUILD_DIR/temp"
 CONFIG_FILE="include/newb_config_legacy.h"
-MANIFEST="$TEMP_PACK_DIR/manifest.json"
-
-# format:
-# pack.sh 15b3 "Custom description if any"
-# requires sed(for editing), zip(for packing)
+PLATFORM=android
 
 BUILD_CODE="14"
 VERSION="[0, $BUILD_CODE, 0]"
 CUSTOM=
-if [ -n "$1" ]; then
-  BUILD_CODE="$1"
-  VER=(${BUILD_CODE:0:2} ${BUILD_CODE:3:5})
-  if [ -n "${VER[1]}" ]; then
-    VERSION="[0, $((${VER[0]}-1)), ${VER[1]}]"
-  else
-    VERSION="[0, ${VER[0]}, 0]"
-  fi
-  if [ -n "$2" ]; then
-    CUSTOM="'$2'"
+ARG_MODE=""
+WIN_BASH=false
+for t in "$@"; do
+  if [ "$t" == "-v" ] || [ "$t" == "-m" ] || [ "$t" == "-p" ]; then
+    # mode
+    ARG_MODE="$t"
+    continue
+  elif [ "$ARG_MODE" == "-v" ]; then
+    # build code
+      BUILD_CODE="$t"
+      VER=(${BUILD_CODE:0:2} ${BUILD_CODE:3:5})
+      if [ -n "${VER[1]}" ]; then
+        VERSION="[0, $((${VER[0]}-1)), ${VER[1]}]"
+      else
+        VERSION="[0, ${VER[0]}, 0]"
+      fi
+  elif [ "$ARG_MODE" == "-m" ]; then
+    # custom message
+    CUSTOM="'$t'"
     BUILD_CODE="${BUILD_CODE}c"
+  elif [ "$ARG_MODE" == "-p" ]; then
+    # platform
+    PLATFORM="$t"
+  elif [ -z "$ARG_MODE" ]; then
+    # using bash on win
+    if [ "$t" == "win" ]; then
+      BUILD_SCRIPT="./build.bat"
+      WIN_BASH=true
+    fi
   fi
-fi
+  ARG_MODE=
+  shift
+done
+
+BUILD_SCRIPT="$BUILD_SCRIPT -p $PLATFORM"
+BUILD_DIR="build/$PLATFORM"
+TEMP_PACK_DIR="$BUILD_DIR/temp"
+MANIFEST="$TEMP_PACK_DIR/manifest.json"
 
 echo -e ">> CREATE PACK FOLDER:\n   - $TEMP_PACK_DIR"
 mkdir -p $TEMP_PACK_DIR
@@ -34,15 +60,20 @@ mkdir -p $TEMP_PACK_DIR/renderer/materials
 cp -rf $PACK_DIR/* $TEMP_PACK_DIR
 
 echo ">> UPDATE MANIFEST"
+if [ "$PLATFORM" == "windows" ]; then
+  sed -i "s/\%w/Only works with BetterRenderDragon/" $MANIFEST
+else
+  sed -i "s/\%w/Only works with Patched Minecraft/" $MANIFEST
+fi
 sed -i "s/\%c/$CUSTOM/" $MANIFEST
 sed -i "s/\"version\": \[.*]/\"version\": $VERSION/g" $MANIFEST
-sed -i "s/\%v-\%p/$BUILD_CODE-android/g" $MANIFEST
-echo "   - code: $BUILD_CODE"
+sed -i "s/\%v-\%p/$BUILD_CODE-$PLATFORM/g" $MANIFEST
+echo "   - code: $BUILD_CODE-$PLATFORM"
 echo "   - version: $VERSION"
 echo "   - custom: $CUSTOM"
 
 echo ">> BUILD MATERIALS"
-$BUILD_SCRIPT -m RenderChunk Clouds Sky EndSky LegacyCubemap Actor
+$BUILD_SCRIPT -m RenderChunk Clouds Sky EndSky LegacyCubemap
 mv -f $BUILD_DIR/*.material.bin $TEMP_PACK_DIR/renderer/materials/
 
 echo ">> BUILD SUBPACKS"
@@ -73,7 +104,9 @@ done
 sed -i "s/\"metadata/\"subpacks\": [\n${CONTENT%,*}\n     ],\n    \"metadata/" $MANIFEST
 sed -i "3s/.*/\/\/ line 3 reserved/" $CONFIG_FILE
 
-cd $TEMP_PACK_DIR
-ZIP_FILE="newb-x-$BUILD_CODE-android.mcpack"
-rm -f $ZIP_FILE
-zip -rq9 ../$ZIP_FILE ./* && echo -e ">> PACKED ZIP\n   - $ZIP_FILE"
+if [ $WIN_BASH = false ]; then
+  cd $TEMP_PACK_DIR
+  ZIP_FILE="newb-x-$BUILD_CODE-$PLATFORM.mcpack"
+  rm -f $ZIP_FILE
+  zip -rq9 ../$ZIP_FILE ./* && echo -e ">> PACKED ZIP\n   - $ZIP_FILE"
+fi
